requirements.txt
numpy
numba
matplotlib
scipy
pyyaml